+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Release Notes
 
+## v1.4.0 / 2022-03-08
+- sync with gitlab
+
+## v1.3.0 / 2020-12-17
+- update trace and errors lib
+
+## v1.2.0 / 2020-12-03
+- add tracing for gentlemanretry
+
 ## v1.1.1 / 2020-11-23
 - migrated to GitHub